# lib/api-zod/src/generated/types/startReasoningBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StartReasoningBody {
  /** When true, begin a fresh attempt even if a previous attempt was already submitted. An in-progress attempt is still resumed. */
  retake?: boolean;
}